제목: feat: 인출 API를 ETH 단위(BigDecimal)로 변경하고 정책·문서 업데이트
본문:

API 변경: CreateWithdrawalRequest 및 관련 컨트롤러에서 amount를 wei(long) 대신 ETH 소수(BigDecimal)로 수신하도록 변경, 내부적으로 ETH→wei 변환 적용
정책: policy.max-amount를 10(ETH)으로 설정하고 PolicyEngine 비교 로직을 BigDecimal로 변경
테스트: 단위/통합 테스트와 JSON 페이로드를 ETH 단위로 수정하여 빌드 통과하도록 조정
문서: README.md를 정리해 기본 모드를 mock으로 지정하고 CUSTODY_CHAIN_MODE로만 모드 제어 안내, Confirmation Tracker 연습 추가
데모: AdapterDemoController에서 ETH 입력 처리 및 변환 추가
기타: idempotency·검증 에러 수정 및 테스트 픽스
파급(중요):
BREAKING CHANGE: 클라이언트/테스트 페이로드는 이제 wei 정수가 아니라 ETH 소수(예: 1 또는 0.5)로 amount를 전송해야 합니다

+++ custody/src/main/java/lab/custody/orchestration/CreateWithdrawalRequest.java
@@ -1,9 +1,11 @@
 package lab.custody.orchestration;
 
+import java.math.BigDecimal;
+
 public record CreateWithdrawalRequest(
         String chainType,
         String fromAddress,
         String toAddress,
         String asset,
-        long amount
+        BigDecimal amount
 ) {}
